@@ -37,28 +37,30 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "7.10.2",
+    "@babel/core": "7.13.0",
     "@babel/plugin-proposal-nullish-coalescing-operator": "7.10.1",
     "@babel/plugin-proposal-optional-chaining": "7.10.1",
     "@babel/preset-typescript": "7.10.1",
     "@hipo/eslint-config-base": "4.1.1",
     "@hipo/eslint-config-react": "2.1.0",
     "@hipo/eslint-config-typescript": "1.0.1",
-    "@hipo/stylelint-config-base": "2.2.0",
-    "@storybook/addon-actions": "6.3.2",
-    "@storybook/addon-links": "6.3.2",
-    "@storybook/addons": "6.3.2",
-    "@storybook/react": "6.3.2",
-    "@testing-library/jest-dom": "5.12.0",
-    "@testing-library/react": "11.2.6",
-    "@testing-library/user-event": "^13.1.9",
+    "@hipo/stylelint-config-base": "3.2.0",
+    "@rollup/plugin-terser": "0.4.0",
+    "@storybook/addon-actions": "6.5.16",
+    "@storybook/addon-links": "6.5.16",
+    "@storybook/addons": "6.5.16",
+    "@storybook/react": "6.5.16",
+    "@storybook/manager-webpack5": "6.5.16",
+    "@testing-library/jest-dom": "5.16.5",
+    "@testing-library/react": "14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "@types/classnames": "2.3.0",
     "@types/jest": "26.0.23",
     "@types/jest-axe": "3.5.1",
-    "@types/react": "17.0.5",
-    "@types/react-dom": "17.0.4",
-    "@types/react-test-renderer": "17.0.1",
-    "@types/uuid": "8.3.4",
+    "@types/react": "18.0.28",
+    "@types/react-dom": "18.0.11",
+    "@types/react-test-renderer": "18.0.0",
+    "@types/uuid": "9.0.1",
     "@typescript-eslint/eslint-plugin": "4.23.0",
     "@typescript-eslint/parser": "4.23.0",
     "axe-core": "4.2.0",
@@ -67,48 +69,48 @@
     "eslint": "7.26.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-import": "2.23.2",
+    "eslint-plugin-jest-dom": "3.9.0",
     "eslint-plugin-jsx-a11y": "6.4.1",
     "eslint-plugin-react": "7.23.2",
     "eslint-plugin-react-hooks": "4.2.0",
-    "eslint-plugin-jest-dom": "3.9.0",
     "eslint-plugin-testing-library": "4.6.0",
     "fork-ts-checker-webpack-plugin": "4.1.6",
-    "husky": "7.0.0",
-    "jest": "26.4.2",
+    "husky": "8.0.3",
+    "jest": "29.0.0",
     "jest-axe": "4.1.0",
-    "lint-staged": "10.2.13",
-    "node-sass": "4.14.1",
-    "prettier": "2.0.5",
+    "lint-staged": "13.1.2",
+    "prettier": "2.8.4",
+    "react": "^18.0.0",
     "react-docgen-typescript-loader": "3.7.2",
+    "react-dom": "^18.0.0",
     "react-svg-loader": "3.0.3",
-    "react-test-renderer": "17.0.2",
-    "rollup": "2.15.0",
+    "react-test-renderer": "18.2.0",
+    "rollup": "3.18.0",
     "rollup-plugin-eslint": "7.0.0",
-    "rollup-plugin-postcss": "3.1.2",
+    "rollup-plugin-postcss": "4.0.2",
     "rollup-plugin-react-svg": "3.0.3",
     "rollup-plugin-stylelint": "1.0.0",
-    "rollup-plugin-terser": "6.1.0",
-    "rollup-plugin-typescript2": "0.27.1",
+    "rollup-plugin-typescript2": "0.34.1",
+    "sass": "1.58.3",
     "sass-loader": "8.0.2",
     "style-loader": "1.2.1",
-    "stylelint": "13.12.0",
-    "stylelint-order": "4.1.0",
-    "stylelint-scss": "3.19.0",
-    "ts-jest": "26.1.0",
+    "stylelint": "14.5.1",
+    "stylelint-order": "5.0.0",
+    "stylelint-scss": "4.0.1",
+    "ts-jest": "29.0.5",
     "ts-loader": "7.0.5",
-    "typescript": "3.9.5",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "typescript": "4.9.5"
   },
   "dependencies": {
-    "classnames": "2.3.1",
-    "date-fns": "2.28.0",
-    "date-fns-tz": "1.3.3",
-    "react-textarea-autosize": "8.3.2",
-    "uuid": "8.3.2"
+    "@storybook/builder-webpack5": "^6.5.16",
+    "classnames": "2.3.2",
+    "date-fns": "2.29.3",
+    "date-fns-tz": "2.0.0",
+    "react-textarea-autosize": "8.4.0",
+    "uuid": "9.0.0"
   },
   "peerDependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react": "^17.0.2 || ^18.0.0",
+    "react-dom": "^17.0.2 || ^18.0.0"
   }
 }
