@@ -62,7 +62,8 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
         autoFocus={shouldFocus}
         onClick={handleClick}
         disabled={isButtonDisabled}
-        {...rest}>
+        {...rest}
+      >
         {children}
 
         {shouldDisplaySpinner && spinnerContent}
