@@ -32,7 +32,8 @@ const FileUploadButton = React.forwardRef<HTMLLabelElement, FileUploadButtonProp
           "button",
           customLabelClassName
         )}
-        onChange={handleFileSelect}>
+        onChange={handleFileSelect}
+      >
         {children}
       </FileInput>
     );
