@@ -1,5 +1,4 @@
 import {axe, toHaveNoViolations} from "jest-axe";
-import {RunOptions} from "axe-core";
 
 expect.extend(toHaveNoViolations);
 
@@ -19,7 +18,7 @@ expect.extend(toHaveNoViolations);
  *
  * @see https://github.com/nickcolley/jest-axe#testing-react-with-react-testing-library
  */
-async function testA11y(container: HTMLElement, options?: RunOptions) {
+async function testA11y(container: HTMLElement, options?: Parameters<typeof axe>[1]) {
   const results = await axe(container, options);
 
   expect(results).toHaveNoViolations();
