@@ -46,18 +46,21 @@ function DateTimer({
     <List
       customClassName={classNames("date-timer", customClassName)}
       testid={testid}
-      items={items}>
+      items={items}
+    >
       {(item, itemTestId) => (
         <ListItem customClassName={"date-timer-box"}>
           <p
             data-testid={`${itemTestId}.${item.id}.count`}
-            className={"date-timer-box__count"}>
+            className={"date-timer-box__count"}
+          >
             {item.count}
           </p>
 
           <p
             data-testid={`${itemTestId}.${item.id}.title`}
-            className={"date-timer-box__title"}>
+            className={"date-timer-box__title"}
+          >
             {item.title}
           </p>
         </ListItem>
