@@ -47,7 +47,8 @@ function FormField(props: FormFieldProps) {
         <List
           testid={`${testid}.error-messages`}
           customClassName={"form-field__error-message-list"}
-          items={errorMessages!}>
+          items={errorMessages!}
+        >
           {(message, messageTestId) => (
             <ListItem customClassName={"form-field__error-message-list__item"}>
               <FormFieldMessage type={"error"} message={message} testid={messageTestId} />
@@ -60,7 +61,8 @@ function FormField(props: FormFieldProps) {
         <List
           testid={`${testid}.helper-messages`}
           customClassName={"form-field__helper-message-list"}
-          items={helperMessages!}>
+          items={helperMessages!}
+        >
           {(message, messageTestId) => (
             <ListItem customClassName={"form-field__helper-message-list__item"}>
               <FormFieldMessage
