@@ -67,7 +67,8 @@ const FileInput = React.forwardRef<HTMLLabelElement, FileInputProps>(
           ref={ref}
           htmlFor={htmlFor}
           className={labelClassName}
-          data-testid={`${testid}.label`}>
+          data-testid={`${testid}.label`}
+        >
           {children}
 
           {isPending && spinnerContent}
