@@ -57,7 +57,8 @@ function PasswordInput({
             aria-label={iconAriaLabel}
             onClick={togglePasswordVisibility}
             shouldStopPropagation={false}
-            shouldPreventDefault={false}>
+            shouldPreventDefault={false}
+          >
             {icon}
           </Button>
         )
