@@ -26,7 +26,8 @@ const DescriptionTerm = React.forwardRef<HTMLDivElement, DescriptionTermProps>(
         className={containerClassName}
         data-testid={testid}
         id={id}
-        role={role}>
+        role={role}
+      >
         <dt className={classNames("description-term__title", titleClassName)}>{title}</dt>
 
         <dd className={classNames("description-term__description", descriptionClassName)}>
