@@ -23,19 +23,16 @@ export interface ProgressBarProps {
 }
 
 function ProgressBar(props: ProgressBarProps) {
-  const {
-    ariaLabelledBy,
-    ariaLabel,
-    title,
-    ariaValueText,
-    ariaDescribedBy,
-    testid
-  } = props;
+  const {ariaLabelledBy, ariaLabel, title, ariaValueText, ariaDescribedBy, testid} =
+    props;
   const {percentage, style, children, customClassName} = props;
   const progressBarClassName = classNames("progress-bar", customClassName);
   const parsedPercentage = percentage >= MAX_VALUE ? MAX_VALUE : percentage;
-  const {trackColor = "blue", backgroundColor = "gray", completedColor = "green"} =
-    style || {};
+  const {
+    trackColor = "blue",
+    backgroundColor = "gray",
+    completedColor = "green"
+  } = style || {};
 
   return (
     <div
@@ -52,7 +49,8 @@ function ProgressBar(props: ProgressBarProps) {
       title={title}
       aria-valuetext={ariaValueText}
       aria-describedby={ariaDescribedBy}
-      data-testid={testid}>
+      data-testid={testid}
+    >
       <div
         className={"progress-bar__track"}
         style={{
