@@ -14,7 +14,8 @@ function SelectGroupComponent(
       ref={ref}
       className={classNames("select-group", customClassName)}
       role={"group"}
-      {...props}>
+      {...props}
+    >
       {children}
     </div>
   );
