@@ -37,7 +37,8 @@ function ToastStack({customRootId}: ToastStackProps) {
         <List
           testid={"ToastStack"}
           items={state.toastStack}
-          customClassName={"toast-stack"}>
+          customClassName={"toast-stack"}
+        >
           {(toast, testid) => <Toast testid={testid} data={toast} />}
         </List>
       ) : (
