@@ -109,7 +109,8 @@ storiesOf("Progress Bar", module)
           backgroundColor: colors.background,
           trackColor: colors.track,
           completedColor: colors.completed
-        }}>
+        }}
+      >
         {<p>{"0 %"}</p>}
       </ProgressBar>
 
@@ -121,7 +122,8 @@ storiesOf("Progress Bar", module)
           backgroundColor: colors.background,
           trackColor: colors.track,
           completedColor: colors.completed
-        }}>
+        }}
+      >
         {<p>{"85 %"}</p>}
       </ProgressBar>
 
@@ -133,7 +135,8 @@ storiesOf("Progress Bar", module)
           backgroundColor: colors.background,
           trackColor: colors.track,
           completedColor: colors.completed
-        }}>
+        }}
+      >
         {<p>{"Completed!"}</p>}
       </ProgressBar>
 
