@@ -58,7 +58,8 @@ storiesOf("Textarea", module).add("Textarea", () => (
     <FormField
       labelledBy={"Address"}
       label={"Address"}
-      errorMessages={["Please enter your address"]}>
+      errorMessages={["Please enter your address"]}
+    >
       <Textarea
         id={"address"}
         name={"Address"}
