@@ -21,7 +21,8 @@ storiesOf("Select", module)
             role={"listbox"}
             options={state.options}
             onSelect={(option) => setState({...state, selectedOption: option})}
-            value={state.selectedOption}>
+            value={state.selectedOption}
+          >
             <Select.Trigger style={{justifyContent: "space-between"}}>
               <div>
                 {state.selectedOption ? state.selectedOption.title : "Select Item"}
@@ -50,7 +51,8 @@ storiesOf("Select", module)
             role={"listbox"}
             options={state.options}
             onSelect={(option) => setState({...state, selectedOption: option})}
-            value={state.selectedOption}>
+            value={state.selectedOption}
+          >
             <Select.Trigger style={{justifyContent: "space-between"}}>
               <div>
                 {state.selectedOption ? state.selectedOption.title : "Select Item"}
@@ -78,7 +80,8 @@ storiesOf("Select", module)
             role={"listbox"}
             options={state.options}
             onSelect={(option) => setState({...state, selectedOption: option})}
-            value={state.selectedOption}>
+            value={state.selectedOption}
+          >
             <Select.Trigger style={{justifyContent: "space-between"}}>
               <div>
                 {state.selectedOption ? state.selectedOption.title : "Select Item"}
@@ -104,7 +107,8 @@ storiesOf("Select", module)
             role={"listbox"}
             options={state.options}
             onSelect={(option) => setState({...state, selectedOption: option})}
-            value={state.selectedOption}>
+            value={state.selectedOption}
+          >
             <Select.Trigger style={{justifyContent: "space-between"}}>
               <div>
                 {state.selectedOption ? state.selectedOption.title : "Select Item"}
@@ -140,7 +144,8 @@ storiesOf("Select", module)
             role={"listbox"}
             options={state.options}
             onSelect={(option) => setState({...state, selectedOption: option})}
-            value={state.selectedOption}>
+            value={state.selectedOption}
+          >
             <Select.Trigger style={{justifyContent: "space-between"}}>
               <div>
                 {state.selectedOption ? state.selectedOption.title : "Select Item"}
@@ -169,7 +174,8 @@ storiesOf("Select", module)
             role={"listbox"}
             options={state.options}
             onSelect={(option) => option && handleMultiSelect(state, setState, option)}
-            value={state.value}>
+            value={state.value}
+          >
             <Select.Trigger style={{justifyContent: "space-between"}}>
               <div>
                 {state.value.length ? `${state.value.length} selected` : "Select Items"}
@@ -201,7 +207,8 @@ storiesOf("Select", module)
             role={"listbox"}
             options={state.options}
             onSelect={(option) => option && handleMultiSelect(state, setState, option)}
-            value={state.value}>
+            value={state.value}
+          >
             <Select.Trigger style={{justifyContent: "space-between"}}>
               <div>
                 {state.value.length ? (
@@ -215,7 +222,8 @@ storiesOf("Select", module)
                               ...state,
                               value: state.value.filter((option) => option.id !== tag.id)
                             })
-                          }></Tag>
+                          }
+                        ></Tag>
                       </ListItem>
                     )}
                   </List>
