@@ -118,7 +118,8 @@ function ClickableUserListItem({user}) {
             onClick: () => {
               setState({displayEmail: !state.displayEmail});
             }
-          }}>
+          }}
+        >
           Click to toggle <em>{user.name}</em>`s email address
           {state.displayEmail && <strong>{` ${user.email}`}</strong>}
         </ListItem>
@@ -156,7 +157,8 @@ storiesOf("List", module)
         placeholderProps={{
           shouldDisplayPlaceholder: true,
           placeholder: renderPlaceholders()
-        }}>
+        }}
+      >
         {(item) => <UserListItem user={item} />}
       </List>
 
@@ -170,7 +172,8 @@ storiesOf("List", module)
         emptyStateProps={{
           shouldDisplayEmptyState: true,
           emptyState: "Sorry, there are no users"
-        }}>
+        }}
+      >
         {(item) => <UserListItem user={item} />}
       </List>
     </Fragment>
